Refactor notices setup preview folder structure
https://eaflood.atlassian.net/browse/WATER-5630

This change moves preview services and presenters into the preview folder.
