+++ app/controllers/notices-setup.controller.js
@@ -39,9 +39,9 @@ const ViewContactTypeService = require('../services/notices/setup/view-contact-t
 const ViewLicenceService = require('../services/notices/setup/view-licence.service.js')
 const ViewNoticeTypeService = require('../services/notices/setup/view-notice-type.service.js')
 const ViewPaperReturnService = require('../services/notices/setup/view-paper-return.service.js')
-const ViewPreviewService = require('../services/notices/setup/view-preview.service.js')
-const ViewPreviewCheckAlertService = require('../services/notices/setup/view-preview-check-alert.service.js')
-const ViewPreviewCheckPaperReturnService = require('../services/notices/setup/view-preview-check-paper-return.service.js')
+const ViewPreviewService = require('../services/notices/setup/preview/view-preview.service.js')
+const ViewPreviewCheckAlertService = require('../services/notices/setup/preview/view-preview-check-alert.service.js')
+const ViewPreviewCheckPaperReturnService = require('../services/notices/setup/preview/view-preview-check-paper-return.service.js')
 const ViewRecipientNameService = require('../services/notices/setup/view-recipient-name.service.js')
 const ViewRemoveLicencesService = require('../services/notices/setup/view-remove-licences.service.js')
 const ViewReturnsPeriodService = require('../services/notices/setup/view-returns-period.service.js')
