+++ app/presenters/notices/setup/preview/preview-check-alert.presenter.js
@@ -5,8 +5,8 @@
  * @module PreviewCheckAlertPresenter
  */
 
-const DetermineRelevantLicenceMonitoringStationsService = require('../../../services/notices/setup/abstraction-alerts/determine-relevant-licence-monitoring-stations.service.js')
-const { determineRestrictionHeading, formatRestrictions } = require('../../monitoring-stations/base.presenter.js')
+const DetermineRelevantLicenceMonitoringStationsService = require('../../../../services/notices/setup/abstraction-alerts/determine-relevant-licence-monitoring-stations.service.js')
+const { determineRestrictionHeading, formatRestrictions } = require('../../../monitoring-stations/base.presenter.js')
 
 /**
  * Formats data for the `/notices/setup/{sessionId}/preview/{contactHashId}/check-alert` page
