+++ app/presenters/notices/setup/preview/preview-check-paper-return.presenter.js
@@ -5,7 +5,7 @@
  * @module PreviewCheckPaperReturnPresenter
  */
 
-const { formatLongDate } = require('../../base.presenter.js')
+const { formatLongDate } = require('../../../base.presenter.js')
 
 /**
  * Formats data for the `/notices/setup/{sessionId}/preview/{contactHashId}/check-paper-return` page
