+++ test/controllers/notices-setup.controller.test.js
@@ -47,9 +47,9 @@ const ViewContactTypeService = require('../../app/services/notices/setup/view-co
 const ViewLicenceService = require('../../app/services/notices/setup/view-licence.service.js')
 const ViewNoticeTypeService = require('../../app/services/notices/setup/view-notice-type.service.js')
 const ViewPaperReturnService = require('../../app/services/notices/setup/view-paper-return.service.js')
-const ViewPreviewService = require('../../app/services/notices/setup/view-preview.service.js')
-const ViewPreviewCheckAlert = require('../../app/services/notices/setup/view-preview-check-alert.service.js')
-const ViewPreviewCheckPaperReturnService = require('../../app/services/notices/setup/view-preview-check-paper-return.service.js')
+const ViewPreviewService = require('../../app/services/notices/setup/preview/view-preview.service.js')
+const ViewPreviewCheckAlert = require('../../app/services/notices/setup/preview/view-preview-check-alert.service.js')
+const ViewPreviewCheckPaperReturnService = require('../../app/services/notices/setup/preview/view-preview-check-paper-return.service.js')
 const ViewRecipientNameService = require('../../app/services/notices/setup/view-recipient-name.service.js')
 const ViewRemoveLicencesService = require('../../app/services/notices/setup/view-remove-licences.service.js')
 const ViewReturnsPeriodService = require('../../app/services/notices/setup/view-returns-period.service.js')
