+++ test/presenters/notices/setup/preview/preview.presenter.test.js
@@ -10,15 +10,15 @@ const { expect } = Code
 
 // Test helpers
 const { HTTP_STATUS_BAD_REQUEST, HTTP_STATUS_OK } = require('node:http2').constants
-const { generateNoticeReferenceCode } = require('../../../../app/lib/general.lib.js')
+const { generateNoticeReferenceCode } = require('../../../../../app/lib/general.lib.js')
 
 // Things we need to stub
-const GeneratePreviewRequest = require('../../../../app/requests/notify/generate-preview.request.js')
+const GeneratePreviewRequest = require('../../../../../app/requests/notify/generate-preview.request.js')
 
 // Thing under test
-const PreviewPresenter = require('../../../../app/presenters/notices/setup/preview.presenter.js')
+const PreviewPresenter = require('../../../../../app/presenters/notices/setup/preview/preview.presenter.js')
 
-describe('Notices - Setup - Preview presenter', () => {
+describe('Notices - Setup - Preview - Preview presenter', () => {
   const contactHashId = '9df5923f179a0ed55c13173c16651ed9'
   const licenceMonitoringStationId = 'a4d15f69-5005-4b6e-ab50-3fbae2deec9c'
   const sessionId = '7334a25e-9723-4732-a6e1-8e30c5f3732e'
